+++ src/graphic/align.h
@@ -58,6 +58,6 @@ inline Align operator|(Align a, Align b) {
 }
 
 Align mirror_alignment(Align alignment);
-void correct_for_align(Align, uint32_t w, uint32_t h, Vector2i* pt);
+void correct_for_align(Align, uint32_t w, uint32_t h, Vector2f* pt);
 }
 #endif  // end of include guard: WL_GRAPHIC_ALIGN_H
