+++ src/graphic/gl/coordinate_conversion.h
@@ -38,7 +38,7 @@ inline void pixel_to_gl_renderbuffer(const int width, const int height, float* x
 // Converts 'rect' given on a screen of 'width' x 'height' pixels into a rect
 // in opengl coordinates in a renderbuffer, i.e. in [-1, 1]. The returned
 // rectangle has positive width and height.
-inline Rectf rect_to_gl_renderbuffer(const int width, const int height, const Recti& rect) {
+inline Rectf rect_to_gl_renderbuffer(const int width, const int height, const Rectf& rect) {
 	float left = rect.x;
 	float top = rect.y;
 	float right = rect.x + rect.w;
