+++ src/graphic/wordwrap.cc
@@ -329,7 +329,7 @@ void WordWrap::draw(RenderTarget& dst, Vector2i where, Align align, uint32_t car
 		if (where.y >= dst.height() || int32_t(where.y + fontheight) <= 0)
 			continue;
 
-		Vector2i point(where.x, where.y);
+		Vector2f point(where.x, where.y);
 
 		if (static_cast<int>(alignment & UI::Align::kRight)) {
 			point.x += wrapwidth_ - LINE_MARGIN;
@@ -346,9 +346,9 @@ void WordWrap::draw(RenderTarget& dst, Vector2i where, Align align, uint32_t car
 			int caret_x = text_width(line_to_caret, style_.font->size());
 
 			const Image* caret_image = g_gr->images().get("images/ui_basic/caret.png");
-			Vector2i caretpt;
+			Vector2f caretpt;
 			caretpt.x = point.x + caret_x - caret_image->width() + LINE_MARGIN;
-			caretpt.y = point.y + (fontheight - caret_image->height()) / 2;
+			caretpt.y = point.y + (fontheight - caret_image->height()) / 2.f;
 			dst.blit(caretpt, caret_image);
 		}
 	}
