+++ src/ui_basic/multilinetextarea.cc
@@ -157,11 +157,11 @@ void MultilineTextarea::draw(RenderTarget& dst) {
 		uint32_t blit_height = std::min(text_im->height(), static_cast<int>(get_inner_h()));
 
 		if (blit_width > 0 && blit_height > 0) {
-			int32_t anchor = 0;
+			float anchor = 0.f;
 			Align alignment = mirror_alignment(align_);
 			switch (alignment & UI::Align::kHorizontal) {
 			case UI::Align::kHCenter:
-				anchor = (get_eff_w() - blit_width) / 2;
+				anchor = (get_eff_w() - blit_width) / 2.f;
 				break;
 			case UI::Align::kRight:
 				anchor = get_eff_w() - blit_width - RICHTEXT_MARGIN;
@@ -170,7 +170,7 @@ void MultilineTextarea::draw(RenderTarget& dst) {
 				anchor = RICHTEXT_MARGIN;
 			}
 
-			dst.blitrect_scale(Recti(anchor, 0, blit_width, blit_height), text_im,
+			dst.blitrect_scale(Rectf(anchor, 0, blit_width, blit_height), text_im,
 			                   Recti(0, scrollbar_.get_scrollpos(), blit_width, blit_height), 1.,
 			                   BlendMode::UseAlpha);
 		}
