+++ src/ui_basic/progresswindow.cc
@@ -57,23 +57,23 @@ ProgressWindow::~ProgressWindow() {
 }
 
 void ProgressWindow::draw_background(RenderTarget& rt, const uint32_t xres, const uint32_t yres) {
-	label_center_.x = xres / 2;
-	label_center_.y = yres * PROGRESS_LABEL_POSITION_Y / 100;
+	label_center_.x = xres / 2.f;
+	label_center_.y = yres * PROGRESS_LABEL_POSITION_Y / 100.f;
 	Recti wnd_rect(Vector2i(0, 0), xres, yres);
 
 	const uint32_t h =
 	   UI::g_fh1->render(as_uifont(UI::g_fh1->fontset()->representative_character()))->height();
 
-	label_rectangle_.x = xres / 4;
-	label_rectangle_.w = xres / 2;
-	label_rectangle_.y = label_center_.y - h / 2 - PROGRESS_STATUS_RECT_PADDING;
-	label_rectangle_.h = h + 2 * PROGRESS_STATUS_RECT_PADDING;
+	label_rectangle_.x = xres / 4.f;
+	label_rectangle_.w = xres / 2.f;
+	label_rectangle_.y = label_center_.y - h / 2.f - PROGRESS_STATUS_RECT_PADDING;
+	label_rectangle_.h = h + 2.f * PROGRESS_STATUS_RECT_PADDING;
 
 	const Image* bg = g_gr->images().get(background_);
-	rt.blitrect_scale(
-	   Recti(0, 0, xres, yres), bg, Recti(0, 0, bg->width(), bg->height()), 1., BlendMode::UseAlpha);
+	rt.blitrect_scale(Rectf(0.f, 0.f, xres, yres), bg, Recti(0, 0, bg->width(), bg->height()), 1.,
+	                  BlendMode::UseAlpha);
 
-	Recti border_rect = label_rectangle_;
+	Rectf border_rect = label_rectangle_;
 	border_rect.x -= PROGRESS_STATUS_BORDER_X;
 	border_rect.y -= PROGRESS_STATUS_BORDER_Y;
 	border_rect.w += 2 * PROGRESS_STATUS_BORDER_X;
